feat: compute highlight overlaps in worker
